+++ test/test-resources.toml
@@ -10,8 +10,8 @@ version = 2
 [[audit_baseline]]
 scope = "all"
 resource = "subprocess"
-baseline_calls = 619
-baseline_files = 178
+baseline_calls = 623
+baseline_files = 179
 reported_calls = 495
 reported_files = 135
 owner_bead = "ga-80po0c.2"
@@ -23,8 +23,8 @@ expires = "2026-10-01"
 [[audit_baseline]]
 scope = "all"
 resource = "fixed_sleep"
-baseline_calls = 452
-baseline_files = 163
+baseline_calls = 453
+baseline_files = 164
 reported_calls = 447
 reported_files = 157
 owner_bead = "ga-80po0c.2"
